refactor: handler unit test, remove secret manager interface
